+++ Tests/Wire/Resources/Shipping/Drivers/ComplianceDocuments/S3Object/ComplianceDocumentsS3ObjectClientWireTests.swift
@@ -8,7 +8,18 @@ import Chrt
         stub.setResponse(
             body: Data(
                 """
-                true
+                {
+                  "_id": "_id",
+                  "blurhash": "blurhash",
+                  "content_type": "content_type",
+                  "driver_compliance_document_id": "driver_compliance_document_id",
+                  "filename": "filename",
+                  "s3_key_prefix": "shipping/driver_compliance_document_s3_object_metadata",
+                  "schema_version": 1,
+                  "uploaded_at_timestamp": "2024-01-15T09:30:00Z",
+                  "uploaded_by_org_id": "uploaded_by_org_id",
+                  "uploaded_by_user_id": "uploaded_by_user_id"
+                }
                 """.utf8
             )
         )
@@ -17,7 +28,18 @@ import Chrt
             token: "<token>",
             urlSession: stub.urlSession
         )
-        let expectedResponse = true
+        let expectedResponse = DriverComplianceDocumentS3ObjectMetadata1(
+            id: "_id",
+            blurhash: Optional("blurhash"),
+            contentType: Optional("content_type"),
+            driverComplianceDocumentId: "driver_compliance_document_id",
+            filename: Optional("filename"),
+            s3KeyPrefix: Optional(.shippingDriverComplianceDocumentS3ObjectMetadata),
+            schemaVersion: 1,
+            uploadedAtTimestamp: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
+            uploadedByOrgId: "uploaded_by_org_id",
+            uploadedByUserId: "uploaded_by_user_id"
+        )
         let response = try await client.shipping.drivers.complianceDocuments.s3Object.addV1(
             driverComplianceDocumentId: "driver_compliance_document_id",
             request: .init(file: .init(data: Data("".utf8))),